+++ app/pb_kits/playbook/kits/pb_badge.js
@@ -1,4 +1,0 @@
-import Badge from '../pb_badge/_badge.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ Badge })