+++ app/pb_kits/playbook/kits/pb_circle_icon_button.js
@@ -1,4 +1,0 @@
-import CircleIconButton from '../pb_circle_icon_button/_circle_icon_button.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ CircleIconButton })