+++ app/pb_kits/playbook/kits/pb_date_range_inline.js
@@ -1,4 +1,0 @@
-import DateRangeInline from '../pb_date_range_inline/_date_range_inline.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ DateRangeInline })