+++ app/pb_kits/playbook/kits/pb_date_stacked.js
@@ -1,4 +1,0 @@
-import DateStacked from '../pb_date_stacked/_date_stacked.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ DateStacked })