+++ app/pb_kits/playbook/kits/pb_fixed_confirmation_toast.js
@@ -1,4 +1,0 @@
-import FixedConfirmationToast from '../pb_fixed_confirmation_toast/_fixed_confirmation_toast.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ FixedConfirmationToast })