+++ app/pb_kits/playbook/kits/pb_flex.js
@@ -1,4 +1,0 @@
-import Flex from '../pb_flex/_flex.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ Flex })