+++ app/pb_kits/playbook/kits/pb_icon.js
@@ -1,4 +1,0 @@
-import Icon from '../pb_icon/_icon.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ Icon })