+++ app/pb_kits/playbook/kits/pb_icon_value.js
@@ -1,4 +1,0 @@
-import IconValue from '../pb_icon_value/_icon_value.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ IconValue })