+++ app/pb_kits/playbook/kits/pb_label_value.js
@@ -1,4 +1,0 @@
-import LabelValue from '../pb_label_value/_label_value.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ LabelValue })