+++ app/pb_kits/playbook/kits/pb_legend.js
@@ -1,4 +1,0 @@
-import Legend from '../pb_legend/_legend.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ Legend })