+++ app/pb_kits/playbook/kits/pb_progress_pills.js
@@ -1,4 +1,0 @@
-import ProgressPills from '../pb_progress_pills/_progress_pills.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ ProgressPills })