+++ app/pb_kits/playbook/kits/pb_selectable_card.js
@@ -1,4 +1,0 @@
-import SelectableCard from '../pb_selectable_card/_selectable_card.jsx'
-
-import WebpackerReact from 'webpacker-react'
-WebpackerReact.setup({ SelectableCard })