+++ app/pb_kits/playbook/kits/pb_text_input.js
@@ -1,4 +1,0 @@
-import TextInput from '../pb_text_input/_text_input.jsx'
-import WebpackerReact from 'webpacker-react'
-
-WebpackerReact.setup({ TextInput })