+++ app/pb_kits/playbook/packs/kits/pb_badge.js
@@ -1,0 +1,1 @@
+import '../../pb_badge/_badge.jsx'