+++ app/pb_kits/playbook/packs/kits/pb_button.js
@@ -1,0 +1,1 @@
+import '../../pb_button/_button.jsx'