+++ app/pb_kits/playbook/packs/kits/pb_dashboard_value.js
@@ -1,0 +1,1 @@
+import '../../pb_dashboard_value/_dashboard_value.jsx'