+++ app/pb_kits/playbook/packs/kits/pb_hashtag.js
@@ -1,0 +1,2 @@
+import '../../pb_hashtag/_hashtag.jsx'
+