+++ app/pb_kits/playbook/packs/kits/pb_icon.js
@@ -1,0 +1,2 @@
+import '../../pb_icon/_icon.jsx'
+