+++ app/pb_kits/playbook/packs/kits/pb_legend.js
@@ -1,0 +1,2 @@
+import '../../pb_legend/_legend.jsx'
+