+++ app/pb_kits/playbook/packs/kits/pb_multiple_users.js
@@ -1,0 +1,2 @@
+import '../../pb_multiple_users/_multiple_users.jsx'
+