+++ app/pb_kits/playbook/packs/kits/pb_multiple_users_stacked.js
@@ -1,0 +1,2 @@
+import '../../pb_multiple_users_stacked/_multiple_users_stacked.jsx'
+