+++ app/pb_kits/playbook/packs/kits/pb_radio.js
@@ -1,0 +1,2 @@
+import '../../pb_radio/_radio.jsx'
+