+++ app/pb_kits/playbook/packs/kits/pb_select.js
@@ -1,0 +1,2 @@
+import '../../pb_select/_select.jsx'
+