+++ app/pb_kits/playbook/packs/kits/pb_user.js
@@ -1,0 +1,2 @@
+import '../../pb_user/_user.jsx'
+