+++ CHANGELOG.md
@@ -10,14 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Flex Kit ([#493][] @markiearnold @kre8sions)
 
+[#493]:https://github.com/powerhome/playbook/pull/493
+
 ### Changed
 
 - Undo Revert for DevDoc Fix [#588](https://github.com/powerhome/playbook/pull/588)
 - Removed popover from examples.js and used MENU constant in doc helper [#591](https://github.com/powerhome/playbook/pull/591)
-[#493]:https://github.com/powerhome/playbook/pull/493
-
-### Changed
 - Fix Form kit data prop merge [#587](https://github.com/powerhome/playbook/pull/587)
+- Fixed WebpackerReact registration [#589](https://github.com/powerhome/playbook/pull/589)
 
 
 ## [4.1.2] 2020-1-30